+++ CHANGELOG.md
@@ -19,6 +19,7 @@ might need to be reworked, and subtyping is obviously no longer supported.
 - `Applicative#lazyZip`, for zipping two applicatives in a way that might not require evaluation of one applicative
 - `MonadT`, a general interface representing monad transformers
 - `MaybeT`, a monad transformer for `Maybe`
+- `Endo`, a monoid formed by `Fn1` under composition
 
 ## [3.3.0] - 2019-02-18
 ### Added